For Eureka Common Server Config
Add a new file to your GitHub repository (the same repository used in the last lab) called “application.yml” (or properties). Because this file is named "application.*", the properties set within apply to all clients of the Config server. This is great for us as we want all clients to find the Eureka server. Add the following key / values (use correct YAML formatting):
eureka.client.serviceUrl.defaultZone=http://localhost:8010/eureka/